web/requirements.txt
Django==1.11.3
django-environ==0.4.3
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.1
requests==2.18.1

